requirements.txt
azure-devops==7.1.0b4
GitPython==3.1.43
pygount==1.6.1
python-dotenv==1.0.0
tabulate==0.9.0
